Renamed status.http.js to statushttp.js
